requirements.txt
Django==3.2.8
djangorestframework==3.12.4
django-cookie-consent==0.2.6
Pillow==8.4.0
python-dotenv==0.19.1
PyYAML==6.0